myproject/requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
simplejwt==2.0.1
sqlparse==0.5.1
typing==3.7.4.3
tzdata==2024.2
whitenoise==6.8.2
